+++ dll/win32/bcrypt/commoncrypto.c
@@ -1,0 +1,155 @@
+#include "precomp.h"
+
+#ifdef HAVE_COMMONCRYPTO_COMMONDIGEST_H
+NTSTATUS process_attach( void *args )
+{
+    return STATUS_SUCCESS;
+}
+
+NTSTATUS process_detach( void *args )
+{
+    return STATUS_SUCCESS;
+}
+
+NTSTATUS hash_init( struct hash *hash )
+{
+    switch (hash->alg_id)
+    {
+    case ALG_ID_MD5:
+        CC_MD5_Init( &hash->u.md5_ctx );
+        break;
+
+    case ALG_ID_SHA1:
+        CC_SHA1_Init( &hash->u.sha1_ctx );
+        break;
+
+    case ALG_ID_SHA256:
+        CC_SHA256_Init( &hash->u.sha256_ctx );
+        break;
+
+    case ALG_ID_SHA384:
+        CC_SHA384_Init( &hash->u.sha512_ctx );
+        break;
+
+    case ALG_ID_SHA512:
+        CC_SHA512_Init( &hash->u.sha512_ctx );
+        break;
+
+    default:
+        ERR( "unhandled id %u\n", hash->alg_id );
+        return STATUS_NOT_IMPLEMENTED;
+    }
+    return STATUS_SUCCESS;
+}
+
+NTSTATUS hmac_init( struct hash *hash, UCHAR *key, ULONG key_size )
+{
+    CCHmacAlgorithm cc_algorithm;
+    switch (hash->alg_id)
+    {
+    case ALG_ID_MD5:
+        cc_algorithm = kCCHmacAlgMD5;
+        break;
+
+    case ALG_ID_SHA1:
+        cc_algorithm = kCCHmacAlgSHA1;
+        break;
+
+    case ALG_ID_SHA256:
+        cc_algorithm = kCCHmacAlgSHA256;
+        break;
+
+    case ALG_ID_SHA384:
+        cc_algorithm = kCCHmacAlgSHA384;
+        break;
+
+    case ALG_ID_SHA512:
+        cc_algorithm = kCCHmacAlgSHA512;
+        break;
+
+    default:
+        ERR( "unhandled id %u\n", hash->alg_id );
+        return STATUS_NOT_IMPLEMENTED;
+    }
+
+    CCHmacInit( &hash->u.hmac_ctx, cc_algorithm, key, key_size );
+    return STATUS_SUCCESS;
+}
+
+
+NTSTATUS hash_update( struct hash *hash, UCHAR *input, ULONG size )
+{
+    switch (hash->alg_id)
+    {
+    case ALG_ID_MD5:
+        CC_MD5_Update( &hash->u.md5_ctx, input, size );
+        break;
+
+    case ALG_ID_SHA1:
+        CC_SHA1_Update( &hash->u.sha1_ctx, input, size );
+        break;
+
+    case ALG_ID_SHA256:
+        CC_SHA256_Update( &hash->u.sha256_ctx, input, size );
+        break;
+
+    case ALG_ID_SHA384:
+        CC_SHA384_Update( &hash->u.sha512_ctx, input, size );
+        break;
+
+    case ALG_ID_SHA512:
+        CC_SHA512_Update( &hash->u.sha512_ctx, input, size );
+        break;
+
+    default:
+        ERR( "unhandled id %u\n", hash->alg_id );
+        return STATUS_NOT_IMPLEMENTED;
+    }
+    return STATUS_SUCCESS;
+}
+
+NTSTATUS hmac_update( struct hash *hash, UCHAR *input, ULONG size )
+{
+    CCHmacUpdate( &hash->u.hmac_ctx, input, size );
+    return STATUS_SUCCESS;
+}
+
+NTSTATUS hash_finish( struct hash *hash, UCHAR *output, ULONG size )
+{
+    switch (hash->alg_id)
+    {
+    case ALG_ID_MD5:
+        CC_MD5_Final( output, &hash->u.md5_ctx );
+        break;
+
+    case ALG_ID_SHA1:
+        CC_SHA1_Final( output, &hash->u.sha1_ctx );
+        break;
+
+    case ALG_ID_SHA256:
+        CC_SHA256_Final( output, &hash->u.sha256_ctx );
+        break;
+
+    case ALG_ID_SHA384:
+        CC_SHA384_Final( output, &hash->u.sha512_ctx );
+        break;
+
+    case ALG_ID_SHA512:
+        CC_SHA512_Final( output, &hash->u.sha512_ctx );
+        break;
+
+    default:
+        ERR( "unhandled id %u\n", hash->alg_id );
+        break;
+    }
+    return STATUS_SUCCESS;
+}
+
+NTSTATUS hmac_finish( struct hash *hash, UCHAR *output, ULONG size )
+{
+    CCHmacFinal( &hash->u.hmac_ctx, output );
+
+    return STATUS_SUCCESS;
+}
+
+#endif /* HAVE_COMMONCRYPTO_COMMONDIGEST_H */
