+++ dll/win32/bcrypt/gnutls.c
@@ -1,0 +1,180 @@
+#include "precomp.h"
+
+#if defined(HAVE_GNUTLS_HASH)
+WINE_DECLARE_DEBUG_CHANNEL(winediag);
+
+static void *libgnutls_handle;
+#define MAKE_FUNCPTR(f) static typeof(f) * p##f
+MAKE_FUNCPTR(gnutls_global_deinit);
+MAKE_FUNCPTR(gnutls_global_init);
+MAKE_FUNCPTR(gnutls_global_set_log_function);
+MAKE_FUNCPTR(gnutls_global_set_log_level);
+MAKE_FUNCPTR(gnutls_hash);
+MAKE_FUNCPTR(gnutls_hash_deinit);
+MAKE_FUNCPTR(gnutls_hash_init);
+MAKE_FUNCPTR(gnutls_hmac);
+MAKE_FUNCPTR(gnutls_hmac_deinit);
+MAKE_FUNCPTR(gnutls_hmac_init);
+MAKE_FUNCPTR(gnutls_perror);
+#undef MAKE_FUNCPTR
+
+static void gnutls_log( int level, const char *msg )
+{
+    TRACE( "<%d> %s", level, msg );
+}
+
+NTSTATUS process_attach( void *args )
+{
+    int ret;
+
+    if (!(libgnutls_handle = wine_dlopen( SONAME_LIBGNUTLS, RTLD_NOW, NULL, 0 )))
+    {
+        ERR_(winediag)( "failed to load libgnutls, no support for crypto hashes\n" );
+        return STATUS_DLL_NOT_FOUND;
+    }
+
+#define LOAD_FUNCPTR(f) \
+    if (!(p##f = wine_dlsym( libgnutls_handle, #f, NULL, 0 ))) \
+    { \
+        ERR( "failed to load %s\n", #f ); \
+        goto fail; \
+    }
+
+    LOAD_FUNCPTR(gnutls_global_deinit)
+    LOAD_FUNCPTR(gnutls_global_init)
+    LOAD_FUNCPTR(gnutls_global_set_log_function)
+    LOAD_FUNCPTR(gnutls_global_set_log_level)
+    LOAD_FUNCPTR(gnutls_hash);
+    LOAD_FUNCPTR(gnutls_hash_deinit);
+    LOAD_FUNCPTR(gnutls_hash_init);
+    LOAD_FUNCPTR(gnutls_hmac);
+    LOAD_FUNCPTR(gnutls_hmac_deinit);
+    LOAD_FUNCPTR(gnutls_hmac_init);
+    LOAD_FUNCPTR(gnutls_perror)
+#undef LOAD_FUNCPTR
+
+    if ((ret = pgnutls_global_init()) != GNUTLS_E_SUCCESS)
+    {
+        pgnutls_perror( ret );
+        goto fail;
+    }
+
+    if (TRACE_ON( bcrypt ))
+    {
+        pgnutls_global_set_log_level( 4 );
+        pgnutls_global_set_log_function( gnutls_log );
+    }
+
+    return STATUS_SUCCESS;
+
+fail:
+    wine_dlclose( libgnutls_handle, NULL, 0 );
+    libgnutls_handle = NULL;
+    return STATUS_DLL_NOT_FOUND;
+}
+
+NTSTATUS process_detach( void *args )
+{
+    pgnutls_global_deinit();
+    wine_dlclose( libgnutls_handle, NULL, 0 );
+    libgnutls_handle = NULL;
+    return STATUS_SUCCESS;
+}
+
+NTSTATUS hash_init( struct hash *hash )
+{
+    gnutls_digest_algorithm_t alg;
+
+    if (!libgnutls_handle) return STATUS_INTERNAL_ERROR;
+
+    switch (hash->alg_id)
+    {
+    case ALG_ID_MD5:
+        alg = GNUTLS_DIG_MD5;
+        break;
+    case ALG_ID_SHA1:
+        alg = GNUTLS_DIG_SHA1;
+        break;
+
+    case ALG_ID_SHA256:
+        alg = GNUTLS_DIG_SHA256;
+        break;
+
+    case ALG_ID_SHA384:
+        alg = GNUTLS_DIG_SHA384;
+        break;
+
+    case ALG_ID_SHA512:
+        alg = GNUTLS_DIG_SHA512;
+        break;
+
+    default:
+        ERR( "unhandled id %u\n", hash->alg_id );
+        return STATUS_NOT_IMPLEMENTED;
+    }
+
+    if (pgnutls_hash_init( &hash->u.hash_handle, alg )) return STATUS_INTERNAL_ERROR;
+    return STATUS_SUCCESS;
+}
+
+NTSTATUS hmac_init( struct hash *hash, UCHAR *key, ULONG key_size )
+{
+    gnutls_mac_algorithm_t alg;
+
+    if (!libgnutls_handle) return STATUS_INTERNAL_ERROR;
+
+    switch (hash->alg_id)
+    {
+    case ALG_ID_MD5:
+        alg = GNUTLS_MAC_MD5;
+        break;
+    case ALG_ID_SHA1:
+        alg = GNUTLS_MAC_SHA1;
+        break;
+
+    case ALG_ID_SHA256:
+        alg = GNUTLS_MAC_SHA256;
+        break;
+
+    case ALG_ID_SHA384:
+        alg = GNUTLS_MAC_SHA384;
+        break;
+
+    case ALG_ID_SHA512:
+        alg = GNUTLS_MAC_SHA512;
+        break;
+
+    default:
+        ERR( "unhandled id %u\n", hash->alg_id );
+        return STATUS_NOT_IMPLEMENTED;
+    }
+
+    if (pgnutls_hmac_init( &hash->u.hmac_handle, alg, key, key_size )) return STATUS_INTERNAL_ERROR;
+    return STATUS_SUCCESS;
+}
+
+NTSTATUS hash_update( struct hash *hash, UCHAR *input, ULONG size )
+{
+    if (pgnutls_hash( hash->u.hash_handle, input, size )) return STATUS_INTERNAL_ERROR;
+    return STATUS_SUCCESS;
+}
+
+NTSTATUS hmac_update( struct hash *hash, UCHAR *input, ULONG size )
+{
+    if (pgnutls_hmac( hash->u.hmac_handle, input, size )) return STATUS_INTERNAL_ERROR;
+    return STATUS_SUCCESS;
+}
+
+NTSTATUS hash_finish( struct hash *hash, UCHAR *output, ULONG size )
+{
+    pgnutls_hash_deinit( hash->u.hash_handle, output );
+    return STATUS_SUCCESS;
+}
+
+NTSTATUS hmac_finish( struct hash *hash, UCHAR *output, ULONG size )
+{
+    pgnutls_hmac_deinit( hash->u.hmac_handle, output );
+    return STATUS_SUCCESS;
+}
+
+#endif /* HAVE_GNUTLS_HASH */
