+++ dll/win32/bcrypt/mbedtls.c
@@ -1,0 +1,249 @@
+#include "precomp.h"
+
+#ifdef SONAME_LIBMBEDTLS
+
+WINE_DEFAULT_DEBUG_CHANNEL(bcrypt);
+
+static void *libmbedtls_handle;
+
+#define MAKE_FUNCPTR(f) static typeof(f) * p##f
+MAKE_FUNCPTR(mbedtls_md_init);
+MAKE_FUNCPTR(mbedtls_md_setup);
+MAKE_FUNCPTR(mbedtls_md_update);
+MAKE_FUNCPTR(mbedtls_md_starts);
+MAKE_FUNCPTR(mbedtls_md_hmac_starts);
+MAKE_FUNCPTR(mbedtls_md_finish);
+MAKE_FUNCPTR(mbedtls_md_hmac_finish);
+MAKE_FUNCPTR(mbedtls_md_free);
+#undef MAKE_FUNCPTR
+
+#define mbedtls_md_init                 pmbedtls_md_init
+#define mbedtls_md_setup                pmbedtls_md_setup
+#define mbedtls_md_update               pmbedtls_md_update
+#define mbedtls_md_starts               pmbedtls_md_starts
+#define mbedtls_md_finish               pmbedtls_md_finish
+#define mbedtls_md_hmac_starts          pmbedtls_md_hmac_starts
+#define mbedtls_md_hmac_finish          pmbedtls_md_hmac_finish
+#define mbedtls_md_free                 pmbedtls_md_free
+
+NTSTATUS process_attach( void *args )
+{
+    if (!(libmbedtls_handle = wine_dlopen( SONAME_LIBMBEDTLS, RTLD_NOW, NULL, 0 )))
+    {
+        ERR( "failed to load libmbedtls, no support for crypto hashes\n" );
+        return STATUS_DLL_NOT_FOUND;
+    }
+
+#define LOAD_FUNCPTR(f) \
+    if (!(p##f = wine_dlsym( libmbedtls_handle, #f, NULL, 0 ))) \
+    { \
+        ERR( "failed to load %s\n", #f ); \
+        goto fail; \
+    }
+
+    LOAD_FUNCPTR(mbedtls_md_init)
+    LOAD_FUNCPTR(mbedtls_md_setup)
+    LOAD_FUNCPTR(mbedtls_md_update)
+    LOAD_FUNCPTR(mbedtls_md_starts)
+    LOAD_FUNCPTR(mbedtls_md_finish)
+    LOAD_FUNCPTR(mbedtls_md_hmac_starts)
+    LOAD_FUNCPTR(mbedtls_md_hmac_finish)
+    LOAD_FUNCPTR(mbedtls_md_free);
+#undef LOAD_FUNCPTR
+
+    return STATUS_SUCCESS;
+
+fail:
+    wine_dlclose( libmbedtls_handle, NULL, 0 );
+    libmbedtls_handle = NULL;
+    return STATUS_DLL_NOT_FOUND;
+}
+
+NTSTATUS process_detach( void *args )
+{
+    if (libmbedtls_handle)
+    {
+        wine_dlclose( libmbedtls_handle, NULL, 0 );
+        libmbedtls_handle = NULL;
+    }
+    return STATUS_SUCCESS;
+}
+
+NTSTATUS hash_init( struct hash *hash )
+{
+#ifndef __REACTOS__
+    if (!libmbedtls_handle) return STATUS_INTERNAL_ERROR;
+#endif
+    switch (hash->alg_id)
+    {
+    case ALG_ID_MD5:
+        mbedtls_md5_init(&hash->u.md5_ctx);
+        mbedtls_md5_starts(&hash->u.md5_ctx);
+        break;
+
+    case ALG_ID_SHA1:
+        mbedtls_sha1_init(&hash->u.sha1_ctx);
+        mbedtls_sha1_starts(&hash->u.sha1_ctx);
+        break;
+
+    case ALG_ID_SHA256:
+        mbedtls_sha256_init(&hash->u.sha256_ctx);
+        mbedtls_sha256_starts(&hash->u.sha256_ctx, FALSE);
+        break;
+
+    case ALG_ID_SHA384:
+    case ALG_ID_SHA512:
+        mbedtls_sha512_init(&hash->u.sha512_ctx);
+        mbedtls_sha512_starts(&hash->u.sha512_ctx, hash->alg_id==ALG_ID_SHA384);
+        break;
+
+    default:
+        ERR( "unhandled id %u\n", hash->alg_id );
+        return STATUS_NOT_IMPLEMENTED;
+    }
+
+    return STATUS_SUCCESS;
+}
+
+NTSTATUS hmac_init( struct hash *hash, UCHAR *key, ULONG key_size )
+{
+    const mbedtls_md_info_t *md_info;
+    mbedtls_md_type_t md_type;
+    int ret;
+#ifndef __REACTOS__
+    if (!libmbedtls_handle) return STATUS_INTERNAL_ERROR;
+#endif
+    mbedtls_md_init(&hash->u.hmac_ctx);
+    switch (hash->alg_id)
+    {
+    case ALG_ID_MD5:
+        md_type = MBEDTLS_MD_MD5;
+        break;
+
+    case ALG_ID_SHA1:
+        md_type = MBEDTLS_MD_SHA1;
+        break;
+
+    case ALG_ID_SHA256:
+        md_type = MBEDTLS_MD_SHA256;
+        break;
+
+    case ALG_ID_SHA384:
+        md_type = MBEDTLS_MD_SHA384;
+        break;
+
+    case ALG_ID_SHA512:
+        md_type = MBEDTLS_MD_SHA512;
+        break;
+
+    default:
+        ERR("unhandled id %u\n", hash->alg_id);
+        return STATUS_NOT_IMPLEMENTED;
+    }
+    if ((md_info = mbedtls_md_info_from_type(md_type)) == NULL)
+    {
+        mbedtls_md_free(&hash->u.hmac_ctx);
+        return STATUS_INTERNAL_ERROR;
+    }
+
+    if ((ret = mbedtls_md_setup(&hash->u.hmac_ctx, md_info, 1)) != 0)
+    {
+        mbedtls_md_free(&hash->u.hmac_ctx);
+        return STATUS_INTERNAL_ERROR;
+    }
+
+    mbedtls_md_hmac_starts(&hash->u.hmac_ctx, key, key_size);
+
+    return STATUS_SUCCESS;
+}
+
+NTSTATUS hash_update( struct hash *hash, UCHAR *input, ULONG size )
+{
+#ifndef __REACTOS__
+    if (!libmbedtls_handle) return STATUS_INTERNAL_ERROR;
+#endif
+    switch (hash->alg_id)
+    {
+    case ALG_ID_MD5:
+        mbedtls_md5_update(&hash->u.md5_ctx, input, size);
+        break;
+
+    case ALG_ID_SHA1:
+        mbedtls_sha1_update(&hash->u.sha1_ctx, input, size);
+        break;
+
+    case ALG_ID_SHA256:
+        mbedtls_sha256_update(&hash->u.sha256_ctx, input, size);
+        break;
+
+    case ALG_ID_SHA384:
+    case ALG_ID_SHA512:
+        mbedtls_sha512_update(&hash->u.sha512_ctx, input, size);
+        break;
+
+    default:
+        ERR( "unhandled id %u\n", hash->alg_id );
+        return STATUS_NOT_IMPLEMENTED;
+    }
+
+    return STATUS_SUCCESS;
+}
+
+NTSTATUS hmac_update( struct hash *hash, UCHAR *input, ULONG size )
+{
+#ifndef __REACTOS__
+    if (!libmbedtls_handle) return STATUS_INTERNAL_ERROR;
+#endif
+    mbedtls_md_update(&hash->u.hmac_ctx, input, size);
+
+    return STATUS_SUCCESS;
+}
+
+NTSTATUS hash_finish( struct hash *hash, UCHAR *output, ULONG size )
+{
+#ifndef __REACTOS__
+    if (!libmbedtls_handle) return STATUS_INTERNAL_ERROR;
+#endif
+    switch (hash->alg_id)
+    {
+    case ALG_ID_MD5:
+        mbedtls_md5_finish(&hash->u.md5_ctx, output);
+        mbedtls_md5_free(&hash->u.md5_ctx);
+        break;
+
+    case ALG_ID_SHA1:
+        mbedtls_sha1_finish(&hash->u.sha1_ctx, output);
+        mbedtls_sha1_free(&hash->u.sha1_ctx);
+        break;
+
+    case ALG_ID_SHA256:
+        mbedtls_sha256_finish(&hash->u.sha256_ctx, output);
+        mbedtls_sha256_free(&hash->u.sha256_ctx);
+        break;
+
+    case ALG_ID_SHA384:
+    case ALG_ID_SHA512:
+        mbedtls_sha512_finish(&hash->u.sha512_ctx, output);
+        mbedtls_sha512_free(&hash->u.sha512_ctx);
+        break;
+
+    default:
+        ERR( "unhandled id %u\n", hash->alg_id );
+        return STATUS_NOT_IMPLEMENTED;
+    }
+
+    return STATUS_SUCCESS;
+}
+
+NTSTATUS hmac_finish( struct hash *hash, UCHAR *output, ULONG size )
+{
+#ifndef __REACTOS__
+    if (!libmbedtls_handle) return STATUS_INTERNAL_ERROR;
+#endif
+    mbedtls_md_hmac_finish(&hash->u.hmac_ctx, output);
+    mbedtls_md_free(&hash->u.hmac_ctx);
+
+    return STATUS_SUCCESS;
+}
+
+#endif /* SONAME_LIBMBEDTLS */
